# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserBilling": {
				"properties": {
					"name": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"region": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"taxNumber": {
						"type": "string",
						"nullable": true
					},
					"registrationNumber": {
						"type": "string",
						"nullable": true
					},
					"bankAccount": {
						"type": "string",
						"nullable": true
					},
					"notes": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name",
					"country",
					"city",
					"address",
					"zipCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_User_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"salt": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/USER_STATUS"
					},
					"enableTwoFactor": {
						"type": "boolean"
					},
					"country": {
						"type": "string"
					},
					"region": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"deletedOn": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double"
					},
					"deletedBy": {
						"type": "number",
						"format": "double"
					},
					"accountVerifyToken": {
						"$ref": "#/components/schemas/AccountVerifyToken"
					},
					"accountVerifyTokenId": {
						"type": "number",
						"format": "double"
					},
					"forgottenPasswordToken": {
						"$ref": "#/components/schemas/ForgottenPasswordToken"
					},
					"forgottenPasswordTokenId": {
						"type": "number",
						"format": "double"
					},
					"twoFactorSession": {
						"$ref": "#/components/schemas/TwoFactorSession"
					},
					"twoFactorSessionId": {
						"type": "number",
						"format": "double"
					},
					"twoFactorAuthentication": {
						"$ref": "#/components/schemas/TwoFactorAuthentication"
					},
					"twoFactorAuthenticationId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"tenant": {
						"$ref": "#/components/schemas/Tenant"
					},
					"tenantId": {
						"type": "number",
						"format": "double"
					},
					"lastLoggedIn": {
						"type": "string",
						"format": "date-time"
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/Document"
						},
						"type": "array"
					},
					"notifications": {
						"$ref": "#/components/schemas/Notifications"
					},
					"entity": {
						"type": "string"
					},
					"enableLogin": {
						"type": "boolean"
					},
					"identifier": {
						"type": "string"
					},
					"employeeType": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"birthName": {
						"type": "string"
					},
					"motherName": {
						"type": "string"
					},
					"placeOfBirth": {
						"type": "string"
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"socialSecurityNumber": {
						"type": "string"
					},
					"taxIdentificationNumber": {
						"type": "string"
					},
					"personalIdentificationNumber": {
						"type": "string"
					},
					"licensePlateNumber": {
						"type": "string"
					},
					"salaries": {
						"items": {
							"$ref": "#/components/schemas/Salary"
						},
						"type": "array"
					},
					"contact": {
						"$ref": "#/components/schemas/Contact"
					},
					"contactId": {
						"type": "number",
						"format": "double"
					},
					"properties": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"billing": {
						"anyOf": [
							{
								"properties": {},
								"type": "object"
							},
							{
								"$ref": "#/components/schemas/UserBilling"
							}
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"CreateUserProperties": {
				"$ref": "#/components/schemas/Omit_User.id-or-createdOn-or-updatedBy-or-updatedOn-or-deletedOn-or-deletedBy-or-accountVerifyToken-or-forgottenPasswordToken_"
			},
...
			"Partial_Document_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/DocumentType"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"ownerType": {
						"$ref": "#/components/schemas/DocumentOwnerType"
					},
					"mimeType": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"properties": {
						"$ref": "#/components/schemas/DocumentProperties"
					},
					"companyId": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"approved": {
						"type": "boolean"
					},
					"stored": {
						"properties": {
							"thumbnail": {
								"type": "string"
							},
							"resized": {
								"type": "string"
							},
							"original": {
								"type": "string"
							}
						},
						"required": [
							"original"
						],
						"type": "object"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Vilkrig x Alpinnovacio CRM API",
		"version": "1.0.0",
		"license": {
			"name": "UNLICENSED"
		},
		"contact": {}
	},
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_User_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of users.\nSecured with JWT token and requires \"User:List\" permission.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwtToken": [
							"User:List",
							"Tenant"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User_"
								}
							}
						}
					}
				},
				"description": "Creates a new user.\nSecured with JWT token and requires \"User:Create\" permission.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwtToken": [
							"User:Create",
							"Tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Properties required to create a user.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserProperties",
								"description": "Properties required to create a user."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUsers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes multiple users by their IDs.\nSecured with JWT token and requires \"User:Delete\" permission.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwtToken": [
							"User:Delete",
							"Tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "An object containing an array of user IDs to delete.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ids": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"ids"
								],
								"type": "object",
								"description": "An object containing an array of user IDs to delete."
							}
						}
					}
				}
			}
		},
...
		"/tenders/{id}/documents/{documentId}": {
			"get": {
				"operationId": "GetTenderDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieves detailed information about a specific tender document by its document ID within a tender.\nThis endpoint is protected by JWT authentication, requiring \"Tender:Get\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Get"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the document to retrieve.",
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"removed": {
											"type": "boolean"
										}
									},
									"required": [
										"removed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Removes a specific document from a tender by its document ID.\nThis endpoint requires JWT authentication and is protected with the \"Tender:Update\" permission,\nensuring that only authorized users can delete documents.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender from which the document will be removed.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the document to be removed.",
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/projects/{id}/documents/{documentId}": {
			"get": {
				"operationId": "GetDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Document_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves detailed information about a specific tender document by its document ID within a tender.\nThis endpoint is protected by JWT authentication, requiring \"Project:GetDocuments\" permission.",
				"tags": [
					"Project"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Project:GetDocuments"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the document to retrieve.",
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"removed": {
											"type": "boolean"
										}
									},
									"required": [
										"removed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Removes a specific document from a tender by its document ID.\nThis endpoint requires JWT authentication and is protected with the \"Tender:DeleteDocument\" permission,\nensuring that only authorized users can delete documents.",
				"tags": [
					"Project"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:DeleteDocument"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender from which the document will be removed.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the document to be removed.",
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/employees": {
			"get": {
				"operationId": "GetEmployees",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_User_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of employees.\nSecured with JWT token and requires \"Employee:List\" permission.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:List",
							"Tenant"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateEmployee",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User_"
								}
							}
						}
					}
				},
				"description": "Creates a new employee.\nSecured with JWT token and requires \"Employee:Create\" permission.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Create",
							"Tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Properties required to create an employee.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserProperties",
								"description": "Properties required to create an employee."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEmployees",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes multiple employees by their IDs.\nSecured with JWT token and requires \"Employee:Delete\" permission.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Delete",
							"Tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "An object containing an array of employee IDs to delete.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ids": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"ids"
								],
								"type": "object",
								"description": "An object containing an array of employee IDs to delete."
							}
						}
					}
				}
			}
		},
		"/employees/{id}": {
			"put": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User_"
								}
							}
						}
					}
				},
				"description": "Updates an existing employee.\nSecured with JWT token and requires \"Employee:Update\" permission.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Update",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Properties to update in the employee.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_User_",
								"description": "Properties to update in the employee."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetEmployee",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User-or-null_"
								}
							}
						}
					}
				},
				"description": "Retrieves a single employee by ID.\nSecured with JWT token and requires \"Employee:Get\" permission.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Get",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes a employee by ID.\nSecured with JWT token and requires \"Employee:Delete\" permission.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Delete",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/employees/{id}/documents": {
			"post": {
				"operationId": "AddDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"uploaded": {
											"type": "boolean"
										}
									},
									"required": [
										"uploaded"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Adds a document to an employee's record specified by the employee ID. This operation is secured with JWT and requires\n\"Employee:Update\" permission, ensuring that only authorized users can add documents.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Update",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the employee to whom the document will be added.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"type": {
										"type": "string"
									},
									"approved": {
										"type": "string"
									},
									"startOfValidity": {
										"type": "string"
									},
									"endOfValidity": {
										"type": "string"
									},
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"type",
									"approved",
									"startOfValidity",
									"endOfValidity",
									"files"
								]
							}
						}
					}
				}
			}
		},
		"/employees/{ownerId}/document/{id}": {
			"delete": {
				"operationId": "RemoveDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"removed": {
											"type": "boolean"
										}
									},
									"required": [
										"removed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Removes a document from an employee's record. This method requires \"Employee:Update\" permission and is secured with JWT,\nallowing only authorized users to delete documents.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Update",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee from whom the document will be removed.",
						"in": "path",
						"name": "ownerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the document to be removed.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDocument",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Document"
								}
							}
						}
					}
				},
				"description": "Updates a specific document for an employee identified by the document ID. This operation is secured with JWT and requires\n\"Employee:Update\" permission, ensuring that only authorized personnel can make changes to documents.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Update",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee whose document is being updated.",
						"in": "path",
						"name": "ownerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the document to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The partial updates to be applied to the document.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Document_",
								"description": "The partial updates to be applied to the document."
							}
						}
					}
				}
			}
		},
		"/employees/{id}/check-documents": {
			"get": {
				"operationId": "CheckDocuments",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Checks the status of documents associated with a specific employee. This method is secured with JWT and requires\n\"Employee:Get\" permission, allowing for verification of document status by authorized users.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Get",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee whose documents are being checked.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/employees/{id}/billing": {
			"put": {
				"operationId": "UpdateBilling",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User_"
								}
							}
						}
					}
				},
				"description": "Updates the billing information for an employee by ID. This operation is secured with JWT and requires\n\"Employee:Update\" permission, ensuring that only authorized users can update billing details.",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"jwtToken": [
							"Employee:Update",
							"Tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the employee whose billing information is being updated.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The billing information to update.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserBilling",
								"description": "The billing information to update."
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}